move screenOptions to hook
